# Lesson49/src/main/java/by/teachmeskills/springbootexample/dto/ProductDto.java
package by.teachmeskills.springbootexample.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@AllArgsConstructor
@NoArgsConstructor
public class ProductDto {
    private int id;
    private String name;
    private String description;
    private int price;
    private int categoryId;
}