# Lesson49/src/main/resources/openapi.json
{"openapi":"3.0.1","info":{"title":"Application API","description":"\"Test OpenAPI Application\"","contact":{"name":"TeachMeSkills","email":"tms@gmail.com"},"license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Dev service"},{"url":"http://localhost:8082","description":"Beta service"}],"tags":[{"name":"product","description":"The product API"},{"name":"category","description":"The Category API"}],"paths":{"/category":{"post":{"tags":["category"],"operationId":"createCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}},"required":true},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}}}},"/category/upload":{"post":{"tags":["category"],"operationId":"createCategory_1","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}}}}}}}},"/product/{id}":{"get":{"tags":["product"],"summary":"Find all products","description":"Find all existed products in Eshop","operationId":"openProductPage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"403":{"description":"Categories not fount - forbidden operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"200":{"description":"All categories were found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}}}}}},"/order/create":{"get":{"tags":["order-controller"],"operationId":"createOrder","responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"200":{"description":"OK"}}}},"/category/single/{id}":{"get":{"tags":["category"],"operationId":"getCategoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}}}},"/category/all":{"get":{"tags":["category"],"summary":"Find all categories","description":"Find all existed categories in Eshop","operationId":"getAllCategories","responses":{"403":{"description":"Categories not fount - forbidden operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"200":{"description":"All categories were found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}}}}}}}}},"components":{"schemas":{"ModelAndView":{"type":"object","properties":{"view":{"$ref":"#/components/schemas/View"},"model":{"type":"object","additionalProperties":{"type":"object"}},"status":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"empty":{"type":"boolean"},"viewName":{"type":"string"},"reference":{"type":"boolean"},"modelMap":{"type":"object","properties":{"empty":{"type":"boolean"}},"additionalProperties":{"type":"object"}}}},"View":{"type":"object","properties":{"contentType":{"type":"string"}}},"CategoryDto":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"rating":{"type":"integer","format":"int32"},"localDate":{"type":"string","format":"date"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}},"ProductDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"integer","format":"int32"},"categoryId":{"type":"integer","format":"int32"}}}}}}